Camera code reorganization into SimpleCamera and Camera (#1354)
* Camera code reorganization into SimpleCamera and Camera
Camera now inherits from SimpleCamera
Fixed bug in Camera.update

* Changed Window.on_resize width and height to integers

* Added SimpleCamera to init __all__.
Sections now use a SimpleCamera instead.
Updated Camera init methods from examples, tests and tutorials to reflect new camera usage

* Changed on_resize argument types to integers

* shake velocity type improved

* Just force a new round as the last one failed on building the docs

+++ arcade/examples/perspective.py
@@ -137,7 +137,7 @@ class Perspective(arcade.Window):
             arcade.set_viewport(0, self.fbo.width, 0, self.fbo.height)
             self.spritelist.draw()
 
-    def on_resize(self, width: float, height: float):
+    def on_resize(self, width: int, height: int):
         super().on_resize(width, height)
         self.program["projection"] = Mat4.perspective_projection(self.aspect_ratio, 0.1, 100, fov=75)
